requirements.txt
google-cloud-speech==2.21.0
pyaudio==0.2.13
websockets==11.0.3
